+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change log for curry-frontend
 Under development
 =================
 
+  * Removed expansion of record types in type error messages
+
   * Replaced MessageM monad with CYT monads and moved CYT monads to curry-base
 
   * Implemented warnings for overlapping module aliases - fixes #14